refactor controls logic react style, now adding mobile contols with nipplejs's virtual joystick
