+++ components/threejs/Buttons3d.tsx
@@ -3,21 +3,41 @@
 import { useContext, useEffect } from "react";
 import { Button } from "../ui/button";
 import { buttonrefctx } from "../Skills";
+import { Dockermodelctx, global3dctx } from "./Root3d";
+import type { global3dctxtypes } from "./Root3d";
+import { lookAtmodel } from "@/lib/three_setup";
+import gsap from "gsap";
 
 const Buttons3d = () => {
   const Enter3dButtonref = useContext(buttonrefctx);
+  const Dockermodel = useContext(Dockermodelctx);
+  const globalcontext: global3dctxtypes | undefined = useContext(global3dctx);
+  // const [ca] = globalcontext
   useEffect(() => {
-    // if (Enter3dButtonref && Enter3dButtonref.current) {
-    //   Enter3dButtonref.current.onclick = () => {
-    //     document.querySelector<HTMLDivElement>("#scene-wrapper")?.focus();
-    //     // window.alert("CLICK");
-    //     gsap.to(Enter3dButtonref.current, {
-    //       opacity: 0,
-    //       duration: 1,
-    //     });
-    //   };
-    // }
-  });
+    if (!globalcontext) return;
+    if (Enter3dButtonref && Enter3dButtonref.current) {
+      Enter3dButtonref.current.onclick = () => {
+        // lookatmodeltl.play();
+        // if (Docekrmodel) {
+        if (!globalcontext.camera || !globalcontext.controls) return;
+        const lookatmodeltl = lookAtmodel({
+          camera: globalcontext.camera,
+
+          controls: globalcontext.controls,
+          model: Dockermodel,
+        });
+        // }
+        // camera.position.set(30.405193262355265, 0, -15.853882753462061);
+
+        // window.alert("testing react stuff");
+        document.querySelector<HTMLDivElement>("#scene-wrapper")?.focus();
+
+        gsap.to(Enter3dButtonref.current, {
+          opacity: 0,
+        });
+      };
+    }
+  }, [Dockermodel]); //dont know why Enter3dButtonref would change but an erro i the future might be coz if this
   return (
     <Button
       id="enter3d"
