+++ lib/three_setup.ts
@@ -368,123 +368,6 @@ export const loadDockerModel = async ({
       data.scene.add(axesHelper);
       // camera.lookAt(new Vector3(10, 10, 0));
 
-      const keysPressed: Record<string, boolean> = {};
-      let speed = 0;
-      let targetSpeed = 0;
-      const baseMaxSpeed = 0.2;
-      const boostedMaxSpeed = 0.5;
-      let maxSpeed = baseMaxSpeed;
-
-      const acceleration = 0.01;
-      const deceleration = 0.01;
-      const brakeDeceleration = 0.5;
-
-      let boosting = false;
-      let braking = false;
-
-      const speedMeter = document.getElementById("speed-meter");
-      const sceneWWrapper = document.getElementById("scene-wrapper");
-
-      sceneWWrapper?.addEventListener("keydown", (ev) => {
-        keysPressed[ev.key] = true;
-        ev.preventDefault();
-        ev.stopPropagation();
-
-        switch (ev.key) {
-          case "ArrowUp":
-            targetSpeed = maxSpeed;
-            break;
-          case "ArrowDown":
-            targetSpeed = -maxSpeed;
-            break;
-          case "Shift":
-            boosting = true;
-            maxSpeed = boostedMaxSpeed;
-            if (keysPressed["ArrowUp"]) targetSpeed = maxSpeed;
-            if (keysPressed["ArrowDown"]) targetSpeed = -maxSpeed;
-            break;
-          case " ":
-            braking = true;
-            break;
-        }
-      });
-
-      sceneWWrapper?.addEventListener("keyup", (ev) => {
-        keysPressed[ev.key] = false;
-        ev.preventDefault();
-        ev.stopPropagation();
-
-        switch (ev.key) {
-          case "ArrowUp":
-          case "ArrowDown":
-            targetSpeed = 0;
-            break;
-          case "Shift":
-            boosting = false;
-            maxSpeed = baseMaxSpeed;
-            if (keysPressed["ArrowUp"]) targetSpeed = maxSpeed;
-            if (keysPressed["ArrowDown"]) targetSpeed = -maxSpeed;
-            break;
-          case " ":
-            braking = false;
-            break;
-        }
-      });
-
-      const gameConrolsLoop = () => {
-        const angle = data.scene.rotation.y;
-
-        // Braking
-        if (braking) {
-          if (speed > 0) {
-            speed = Math.max(speed - brakeDeceleration, 0);
-          } else if (speed < 0) {
-            speed = Math.min(speed + brakeDeceleration, 0);
-          }
-        } else {
-          // Normal acceleration/deceleration
-          if (speed < targetSpeed) {
-            speed = Math.min(speed + acceleration, targetSpeed);
-          } else if (speed > targetSpeed) {
-            speed = Math.max(speed - deceleration, targetSpeed);
-          }
-        }
-
-        // Move scene
-        if (Math.abs(speed) > 0.001) {
-          data.scene.position.x -= Math.cos(-angle) * speed;
-          data.scene.position.z -= Math.sin(-angle) * speed;
-        }
-
-        if (keysPressed["ArrowRight"]) {
-          data.scene.rotation.y -= degToRad(1);
-          if (!ships_wheel) return;
-
-          //TODO
-
-          // ships_wheel.rotation.x -= degToRad(1);
-          // pivot.rotateX(-degToRad(1)); t
-        }
-        if (keysPressed["ArrowLeft"]) {
-          data.scene.rotation.y += degToRad(1);
-          if (!ships_wheel) return;
-          // ships_wheel.rotation.x += degToRad(1);
-          // pivot.rotateX(degToRad(1));
-        }
-
-        // Update speed meter
-        // if (speedMeter) {
-        //   const percentage = Math.round(
-        //     (Math.abs(speed) / boostedMaxSpeed) * 100,
-        //   );
-        //   speedMeter.innerText = `Speed: ${percentage}knots`;
-        // }
-
-        requestAnimationFrame(gameConrolsLoop);
-      };
-
-      gameConrolsLoop();
-
       //there is overlap of the timeout in the gamloop below
       //  type gamekeys = {
       //       ArrowUp?: boolean;
@@ -953,7 +836,7 @@ function ModelAnimations(
       controlsTarget2,
       {
         x_v: 10.405193262355265,
-        y_v: 5,
+        y_v: 3,
         z_v: -15.853882753462061,
         ease: "none",
         duration: 2,
